requirements.txt
pyopenssl
urllib3
requests
selenium
tqdm
progressbar
